backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
